DI-2892: Renamed top package from ec2 to vrvm_new
