+++ spot-pricing/vrvm_new/pricing.py
